refactor(cli): document types
