+++ packages/core/src/generators/stencil/generate.ts
@@ -14,7 +14,7 @@ import {
 } from '../../modules/plugins';
 import { fastClone } from '../../helpers/fast-clone';
 import { stripMetaProperties } from '../../helpers/strip-meta-properties';
-import { BaseTranspilerOptions, Transpiler } from '../../types/config';
+import { BaseTranspilerOptions, Transpiler } from '../../types/transpiler';
 import { collectClassString } from './collect-class-string';
 import { getProps } from '../../helpers/get-props';
 import { stripStateAndPropsRefs } from '../../helpers/strip-state-and-props-refs';
