+++ packages/core/src/index.ts
@@ -64,6 +64,7 @@ export * from './helpers/is-mitosis-node';
 export * from './types/mitosis-node';
 export * from './types/mitosis-component';
 export * from './types/config';
+export * from './types/transpiler';
 
 export * from './plugins/compile-away-builder-components';
 export * from './plugins/compile-away-components';
