+++ packages/core/src/types/config.ts
@@ -1,6 +1,3 @@
-import { MitosisComponent } from '..';
-import { Plugin } from './plugins';
-
 export type Format = 'esm' | 'cjs';
 interface TranspilerOptions {
   format?: Format;
@@ -14,24 +11,30 @@ export type GeneratorOptions = {
   };
 };
 
-export interface TranspilerArgs {
-  path?: string;
-  component: MitosisComponent;
-}
-
-export type Transpiler = (args: TranspilerArgs) => string;
-
-export interface BaseTranspilerOptions {
-  experimental?: { [key: string]: any };
-  prettier?: boolean;
-  plugins?: Plugin[];
-}
-
 export type MitosisConfig = {
-  type?: 'library'; // Only one type right now
+  /**
+   * List of targets to compile to.
+   */
   targets: Target[];
+  /**
+   * The output directory. Defaults to `output`.
+   */
   dest?: string;
+  /**
+   * globs of files to transpile. Defaults to `src/*`.
+   */
   files?: string | string[];
+  /**
+   * The directory where overrides are stored. The structure of the override directory must match that of the source code,
+   * with each target having its own sub-directory: `${overridesDir}/${target}/*`
+   * Defaults to `overrides`.
+   */
   overridesDir?: string;
+  /**
+   * Options for each transpilers, e.g.
+   * ```json
+   *
+   * ```
+   */
   options: Partial<GeneratorOptions>;
 };
